requirements.txt
amqp==5.0.2
asgiref==3.2.3
billiard==3.6.3.0
CacheControl==0.12.6
cachetools==4.0.0
celery==5.0.2
certifi==2019.11.28
chardet==3.0.4
click==7.1.2
click-didyoumean==0.0.3
click-repl==0.1.6
configparser==5.0.1
coreapi==2.3.3
coreschema==0.0.4
defusedxml==0.6.0
diff-match-patch==20181111
dj-database-url==0.5.0
Django==2.2.9
django-2gis-maps==1.0.2
django-ckeditor==5.8.0
django-cors-headers==3.2.1
django-debug-toolbar==2.2
django-filter==2.2.0
django-import-export==2.2.0
django-jet==1.0.8
django-js-asset==1.2.2
django-minio-storage==0.3.10
django-mptt==0.11.0
django-rest-framework==0.1.0
django-widget-tweaks==1.4.5
djangorestframework==3.11.0
djangorestframework-camel-case==1.1.2
drf-yasg==1.17.1
et-xmlfile==1.0.1
factory-boy==2.12.0
Faker==4.0.1
firebase-admin==2.17.0
google-api-core==1.16.0
google-api-python-client==1.7.11
google-auth==1.11.2
google-auth-httplib2==0.0.3
google-cloud-core==1.3.0
google-cloud-firestore==1.6.2
google-cloud-storage==1.26.0
google-resumable-media==0.5.0
googleapis-common-protos==1.51.0
grpcio==1.27.2
gunicorn==20.0.4
httplib2==0.17.0
idna==2.8
inflection==0.3.1
itypes==1.1.0
jdcal==1.4.1
Jinja2==2.11.1
kombu==5.0.2
Markdown==3.1.1
MarkupPy==1.14
MarkupSafe==1.1.1
minio==6.0.2
msgpack==1.0.0
odfpy==1.4.1
openpyxl==3.0.4
packaging==20.3
Pillow==7.0.0
prompt-toolkit==3.0.8
protobuf==3.11.3
psycopg2-binary==2.8.5
pyasn1==0.4.8
pyasn1-modules==0.2.8
pyparsing==2.4.6
python-dateutil==2.8.1
pytils==0.3
pytz==2019.3
requests==2.22.0
rsa==4.0
ruamel.yaml==0.16.12
ruamel.yaml.clib==0.2.2
six==1.14.0
sqlparse==0.3.0
tablib==2.0.0
text-unidecode==1.3
uritemplate==3.0.1
urllib3==1.25.8
vine==5.0.0
wcwidth==0.2.5
xlrd==1.2.0
xlwt==1.3.0
